+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/gateway
 
+## 1.1.0-beta.17
+
+### Patch Changes
+
+- 1d8ea2c: feat(provider/gateway): Add GPT-5 pro to Gateway model string autocomplete
+
+## 1.1.0-beta.16
+
+### Patch Changes
+
+- Updated dependencies [046aa3b]
+- Updated dependencies [e8109d3]
+  - @ai-sdk/provider@2.1.0-beta.5
+  - @ai-sdk/provider-utils@3.1.0-beta.7
+
 ## 1.1.0-beta.15
 
 ### Patch Changes