+++ packages/google/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/google
 
+## 2.1.0-beta.11
+
+### Patch Changes
+
+- 9a728c8: support latest gemini model id
+
+## 2.1.0-beta.10
+
+### Patch Changes
+
+- e8109d3: feat: tool execution approval
+- Updated dependencies [046aa3b]
+- Updated dependencies [e8109d3]
+  - @ai-sdk/provider@2.1.0-beta.5
+  - @ai-sdk/provider-utils@3.1.0-beta.7
+
 ## 2.1.0-beta.9
 
 ### Patch Changes